create crate compiler-base and move new_error into it
