+++ kclvm/3rdparty/rustc_error/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod snippet;
