+++ kclvm/compiler_base/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "compiler-base"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+[workspace]
+members = [
+    # "diagnostic",
+    "error",
+    "macros",
+    # "style"
+]
