+++ kclvm/compiler_base/diagnostic/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "diagnostic"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+style = {path="../style", version="0.1.0"}
+kclvm-span = {path = "../../span", version = "0.1.0"}
+macros = {path = "../macros", version = "0.1.0"}
+
+termcolor = "1.0"
+style-term = "1.0"
+
