+++ kclvm/compiler_base/error/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "error"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+diagnostic = {path = "../diagnostic", version = "0.1.0"}
+macros = {path = "../macros", version = "0.1.0"}
