+++ kclvm/compiler_base/style/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "style"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+termcolor = "1.0"
