+++ kclvm/compiler_base/style/src/diagnostic/mod.rs
@@ -1,0 +1,1 @@
+pub mod shader;
