+++ kclvm/compiler_base/style/src/tests.rs
@@ -1,0 +1,3 @@
+// shader
+#[test]
+fn test_DiagnosticStyle() {}
