--- kclvm/new_err_macro/src/lib.rs
@@ -1,21 +1,0 @@
-use diagnostics::diagnostic_derive;
-use synstructure::decl_derive;
-
-mod diagnostics;
-
-#[cfg(test)]
-mod tests;
-
-decl_derive!(
-    [DiagnosticBuilder, attributes(
-        // struct attributes
-        warning,
-        error,
-        note,
-        help,
-        position,
-        nopendant,
-        title,
-        // field attributes
-        )] => diagnostic_derive
-);
