--- kclvm/new_err_msg/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "new_err_msg"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-fluent-bundle = "0.15.2"
-fluent-syntax = "0.11"
-intl-memoizer = "0.5.1"
-unic-langid = { version = "0.9.0", features = ["macros"] }
