--- kclvm/new_err_msg/locales/en-US/borrowck.ftl
@@ -1,18 +1,0 @@
-borrowck-move-unsized =
-    cannot move a value of type `{$ty}`
-    .label = the size of `{$ty}` cannot be statically determined
-
-borrowck-higher-ranked-lifetime-error =
-    higher-ranked lifetime error
-
-borrowck-could-not-prove =
-    could not prove `{$predicate}`
-
-borrowck-could-not-normalize =
-    could not normalize `{$value}`
-
-borrowck-higher-ranked-subtype-error =
-    higher-ranked subtype error
-  
-generic-does-not-live-long-enough =
-    `{$kind}` does not live long enough
