--- kclvm/new_err_msg/locales/en-US/builtin_macros.ftl
@@ -1,5 +1,0 @@
-builtin_macros-requires-cfg-pattern =
-    macro requires a cfg-pattern as an argument
-    .label = cfg-pattern required
-
-builtin_macros-expected-one-cfg-pattern = expected 1 cfg-pattern
