--- kclvm/new_error/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "new_error"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-kclvm-error = {path = "../error"}
-kclvm-span = {path = "../span", version = "0.1.0"}
-rustc_span = { path = "../3rdparty/rustc_span" }
-new_err_macro = {path = "../new_err_macro", version = "0.1.0"}
-
-termcolor = "1.0"
-style-term = "1.0"
-
