--- kclvm/new_error/src/lib.rs
@@ -1,12 +1,0 @@
-// mod formatter;
-pub mod diagnostic;
-pub mod emitter;
-pub mod error;
-pub mod pendant;
-pub mod sentence;
-pub mod shader;
-pub mod snippet;
-pub mod styled_buffer;
-
-#[cfg(test)]
-mod tests;
